+++ HISTORY.md
@@ -7,6 +7,8 @@
 - Implemented new method `toNumeric` for units.
 - Implemented new units `arcsec`, `arcsecond`, `arcmin`, `arcminute`. 
   Thanks @devdevdata222.
+- Fixed #485: Scoping issue with variables both used globally as well as in a 
+  function definition.
 - Fixed: Function `number` didn't support `Fraction` as input.